fixed linting, removed image api, added documentation, fixed package.json
