--- .storybook/index.tsx
@@ -1,7 +1,0 @@
-import { getStorybookUI } from '@storybook/react-native';
-
-import './storybook.requires';
-
-const StorybookUIRoot = getStorybookUI({});
-
-export default StorybookUIRoot;
