+++ App.ts
@@ -1,8 +1,9 @@
+// Export main app
 import { App } from './src';
-import StorybookUIRoot from './.ondevice/Storybook';
 
-// const STORYBOOK_START = true;
-console.log(process.env.STORYBOOK_START);
-export default process.env.STORYBOOK_START ? StorybookUIRoot : App;
-// export default App;
+export default App;
+
+// Export storybook mobile UI -- you can still run 'yarn storybook' for web
+// import StorybookUIRoot from './.ondevice/Storybook';
+
 // export default StorybookUIRoot;
