+++ src/services/user-api.ts
@@ -1,5 +1,6 @@
-import { getUsers } from 'src/firebase/user-api';
+import { getUsers, fetchUserImage } from 'src/firebase/user-api';
 import { PublicUserData } from 'src/types';
+import { ImageOBJ } from 'src/types/profile-image';
 import { ConfigApi } from './config-api';
 
 /**
@@ -32,8 +33,27 @@ export const UserApi = ConfigApi.injectEndpoints({
                 }
             },
         }),
+        getUserImage: build.query<string | null, ImageOBJ>({
+            /**
+             * Fetches the usser image either from firestore or uploads new image from state change.
+             *
+             * @param {*} obj - takes in an object with with userID(required) and iamgeUri(optional --upload only)
+             * @return {*} - returns a string or null
+             */
+            async queryFn(obj) {
+                try {
+                    const image = await fetchUserImage(obj.userID, obj.imageUri);
+                    // set into firebase object
+                    console.warn('image data being sent back', image);
+                    return { data: image };
+                } catch (e: any) {
+                    console.warn(`Error with fetching users`);
+                    return { error: e };
+                }
+            },
+        }),
     }),
     overrideExisting: true,
 });
 
-export const { useGetUsersQuery } = UserApi;
+export const { useGetUsersQuery, useGetUserImageQuery } = UserApi;
