+++ src/utils/upload-image.ts
@@ -2,10 +2,15 @@ import * as ImagePicker from 'expo-image-picker';
 import React from 'react';
 import { getDownloadURL, ref, uploadBytesResumable } from 'firebase/storage';
 import { storage, db } from 'src/firebase/config';
-
 import { updateDoc, doc } from 'firebase/firestore';
 
-export const takePhoto = async (
+export /**
+ * Lauches expo image picker for camera and changes the imageState to the URI
+ *
+ * @param {(React.Dispatch<React.SetStateAction<string | undefined>>)} setImageState
+ */
+
+const takePhoto = async (
     setImageState: React.Dispatch<React.SetStateAction<string | undefined>>,
 ) => {
     const pickerResult = await ImagePicker.launchCameraAsync({
@@ -14,16 +19,17 @@ export const takePhoto = async (
     });
 
     if (!pickerResult.cancelled) {
-        // console.log(pickerResult.uri);
-
-        // const img = await fetch(pickerResult.uri);
-        // const bytes = await img.blob();
-
         setImageState(pickerResult.uri);
     }
 };
 
-export const pickImage = async (
+export /**
+ * Lauches expo image picker for camera roll and changes the imageState to the URI
+ *
+ * @param {(React.Dispatch<React.SetStateAction<string | undefined>>)} setImageState
+ */
+
+const pickImage = async (
     setImageState: React.Dispatch<React.SetStateAction<string | undefined>>,
 ) => {
     const result = await ImagePicker.launchImageLibraryAsync({
@@ -33,15 +39,20 @@ export const pickImage = async (
         quality: 1,
     });
 
-    // console.log(result);
-
     if (!result.cancelled) {
-        // console.log(result.uri);
         setImageState(result.uri);
     }
 };
 
-export const upLoadFile = async (file: string, userID: string): Promise<void> => {
+export /**
+ * Uploads file to firebase storage
+ *
+ * @param {string} file
+ * @param {string} userID
+ * @return {*}  {Promise<void>}
+ */
+
+const upLoadFile = async (file: string, userID: string): Promise<void> => {
     if (!file) return;
     const img = await fetch(file);
     const blobFile = await img.blob();
@@ -50,18 +61,16 @@ export const upLoadFile = async (file: string, userID: string): Promise<void> =>
 
     const userRef = doc(db, 'private-user-data', userID);
 
-    console.log('file from upLoadFile', file);
-
     uploadImage.on(
         'state_changed',
         (snapshot) => {
             const percent = Math.round((snapshot.bytesTransferred / snapshot.totalBytes) * 100);
+            // useful log for seeing image being uploaded
             console.log(percent);
         },
-        (err) => console.log(err),
+        (err) => console.warn(err),
         () => {
             getDownloadURL(uploadImage.snapshot.ref).then(async (url) => {
-                console.log(url);
                 const data = {
                     image: url,
                 };
